Update streak_StreamlabsSystem.py
Added a dec function, which when called will decrease the BEST AND ACTIVE streak on the given killer, so make sure you !streak add back to that killer if need be! Other than that, added a message that displays when a non-mod tries to use a mod only command (Sorry, this is for mods only!) and a message if someone tries to do "!streak <invalid input>" (The official uses are '!streak view' or '!streak view <killer>', so try those!). Added a message if an invalid killer name is detected in !streak add, !streak set, !streak reset, and !streak dec. Also added some error messages that *should* never show up, and if they do, they'll tell you to tell me ASAP. I probably also did some more minor changes that I forget to document. If there's an issue, let me know ASAP and I'll fix it!